Silence GCC warning

+++ src/lv_core/lv_debug.h
@@ -49,7 +49,7 @@ void lv_debug_log_error(const char * msg, uint64_t value);
 do {                                            \
     if(!(expr)) {                               \
         LV_LOG_ERROR(__func__);                 \
-        lv_debug_log_error(msg, (uint64_t)value);         \
+        lv_debug_log_error(msg, (uint64_t)((uintptr_t)value));         \
         while(1);                               \
     }                                           \
 } while(0)
